Added score to GameManager, began some refactoring for the GUI scripting... read more:
My goal is to have 1 place for all the GUI scripting (or at least by far
most of it, e.g. all the parts that are fixed on screen) instead of
little pieces everywhere.
Thus, I will subsequently reorganize some references of the player
script to the GUI script, and maybe move some variables from the Player
to the GameManager script.

And I (hopefully) fixed a little bug with the K cheat that caused a
NullReferenceException before...