PLEASE READ, IMPORTANT BUG NOTICE:
ADDED: some classes that should be taking care of an AI. State is an abstract base class implementing the run function, combat a simple interface that just makes sure certain classes must have a shoot function (not realy necessary actually at this point).
ATTENTION: At the moment there is a bug where in the AI.cs the derived classes always give back/pass "null" ?!?! Maybe I am missing something that is specific to C# that I do not know about. Perhaps someone else knows instantly what the problem is when looking at it ?!, but when that works the simple sample States (just doint nothing or stopping a bit away from the player and shooting at him) should be fine.

-Enemies need audio source to have sound when shooting